Since the tun/tap code uses `yggdrasil.Conn` objects without setting deadline, I modified the way `Read` and `Write` are implemented to avoid allocating a `util.Cancellation` in cases where no deadline is set.

This speeds things up a bit regardless of MTU, but is (proportionally) more significant for small MTU:
mtu=1280: 480 Mbps -> 550 Mbps
mtu=65535: 4.3 Gbps -> 4.5 Gbps